requirements.txt
matplotlib
pandas
seaborn
plotly
dill
pyarrow
pytest
pytest-mock
python-dateutil
starlette
requests
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3